# src/main/resources/static/api-docs-swagger.yaml
openapi: 3.0.2

info:
  title: Minesweeper API
  description: API used for the MineSweeper game
  version: 1.0.0
paths:

  /games:
    post:
      tags:
        - Game
      summary: Creates a new MineSweeper game
      description: Creates a new MineSweeper game
      operationId: createGame
      responses:
        201:
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedGameResponse"
              example:
                gameId: 3160c9de-b152-4886-ae52-41f670c493e9
        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"

  /games/{gameId}/cells/{cellId}/reveal:
    post:
      tags:
        - Game
      summary: Reveal the content of a cell
      description: Reveal the content of a cell in a MineSweeper game
      operationId: revealCell
      parameters:
        - name: gameId
          in: path
          description: Id of the minesweeper game
          required: true
          schema:
            type: integer
        - name: cellId
          in: path
          description: Id of the cell in a minesweeper game
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Cell(s) revealed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellReveledResponse"
        404:
          description: CellId or GameId not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
  /games/{gameId}/cells/{cellId}/mark:
    post:
      tags:
        - Game
      summary: Marks a cell
      description: Marks a cell, changing it status in a MineSweeper game
      operationId: markCell
      parameters:
        - name: gameId
          in: path
          description: Id of the minesweeper game
          required: true
          schema:
            type: integer
        - name: cellId
          in: path
          description: Id of the cell in a minesweeper game
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Cell marked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkCellResponse"
        404:
          description: CellId or GameId not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
components:
  schemas:
    MarkCellRequest:
      type: object
      properties:
        currentCellStatus:
          type: string
          description: Current cell status
          example: "UNCHECKED"
    MarkCellResponse:
      type: object
      properties:
        currentCellStatus:
          type: string
          description: Current cell status
          example: "FLAGGED"
    CellReveledResponse:
      type: object
      properties:
        revealedCells:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ReveledCells'
          example:
            10: "1"
            33: "4"
    ReveledCells:
      type: object
      properties:
        id:
          type: string
        value:
          type: string
    CreatedGameResponse:
      type: object
      properties:
        gameId:
          type: string
          description: Game id
          example: "3160c9de-b152-4886-ae52-41f670c493e9"
    ResourceNotFoundResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: timestamp of the response
          example: "2020-05-30T21:09:11.897+0000"
        status:
          type: integer
          description: http status code
          example: 404
        error:
          type: string
          description: Description of the error
          example: "Resource was not found"
        message:
          type: string
          description: Message of the response
          example: "CellId was not found"
        path:
          type: string
          description: Path called
          example: "/games/3160c9de-b152-4886-ae52-41f670c493e9/cells/99"

    MethodNotAllowedErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: timestamp of the response
          example: "2020-05-30T21:09:11.897+0000"
        status:
          type: integer
          description: http status code
          example: 405
        error:
          type: string
          description: Description of the error
          example: "Method Not Allowed"
        message:
          type: string
          description: Message of the response
          example: ""
        path:
          type: string
          description: Path called
          example: "/endpoint"